requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
youtube-transcript-api>=0.6.0
streamlit>=1.30.0
